refactor(test): extract shared MongoDB and encryption-key helpers into test_utils
